+++ relay-server/src/actors/store.rs
@@ -199,6 +199,7 @@ impl StoreService {
                     scoping.organization_id,
                     scoping.project_id,
                     start_time,
+                    client,
                     retention,
                     item,
                 )?,
@@ -783,6 +784,7 @@ impl StoreService {
         organization_id: u64,
         project_id: ProjectId,
         start_time: Instant,
+        client: Option<&str>,
         retention_days: u16,
         item: &Item,
     ) -> Result<(), StoreError> {
@@ -790,6 +792,7 @@ impl StoreService {
             project_id,
             retention_days,
             start_time: UnixTimestamp::from_instant(start_time).as_secs(),
+            sdk: client.map(str::to_owned),
             payload: item.payload(),
         });
 
@@ -1052,6 +1055,8 @@ struct CheckInKafkaMessage {
     payload: Bytes,
     /// Time at which the event was received by Relay.
     start_time: u64,
+    /// The SDK client which produced the event.
+    sdk: Option<String>,
     /// The project id for the current event.
     project_id: ProjectId,
     // Number of days to retain.